Renamed hazardlib.sec_perils -> sep.classes
